fix bug (handling of alternate forms with no default form)
When only "alternate forms", and not a "default form", were specified for a property in order to generate red links, too many error messages where displayed.
Basically the reporting level was not accounted for and the fix consists of only reporting warnings and errors (i.e. no notices and no debug messages).

In addition alternate forms where not taken into account when looking for the form to use.

Now the following logic is used to find a form:
* Try for a form explicitly set as a URL parameter
* Try for an alternate form explicitly set as a URL parameter
* Try for a default form for the given target page
* Try for a default alternate form for the given target page
* Throw an error

Change-Id: Ibcfdee695d4498cfcbaf28bdc1a7d78dc88bbee7