fix bug 43989 (Ampersands not handled properly in SF tags)
The problem is, that the Parser when parsing the form description replaces all text (including anything in <nowiki> tags) by its
HTML representation. This means, <nowiki> tags are not good enough to protect SF tags from the parser.

The solution is to use proper strip items.

It is not as nice as it should be, but that is mainly because of MediaWiki design issues:
* insertStripItem() should be a method of the StripItem class, not of Parser
* StripItem should have a getter method for the marker prefix

Bug: 43989
Change-Id: I8be8f5ce7fd091873eb816004db1a2d69858304e