Use JS from OpenLayers extension if it is installed
This introduces optional dependacys on OpenLayers mediawiki extension.

If that extension is installed openlayers map will use the new updated
maps js code if the extension is not installed it will fallback and use
the website.

The reason for two files that are the same but one is updated is because
the way they were updated stoped it working with the openlayers website
but only worked with it if you had the openlayers js file. So by doing
this way you can stay backword compatibility by making OpenLayers
extension a Dependacy optional.

Change-Id: Id698203f9cb2ddbef903ae924223b1bf34eab566