Add a 'returnto' parameter
Add a 'returnto' parameter to be used in #forminput and #formlink parser functions
or manually constructed links to forms

Sometimes you want to edit a page and return to a different one after saving. Imagine
you have a list of contacts on an overview page. The new 'returnto' parameter allows
you to link to edit forms for every entry and to create forms for new entries and bring
the user directly back to the overview page after hitting the 'save' button.

Usage example:

{{#formlink:form=SampleForm|link text=come back here|query string=returnto={{PAGENAME}}}}

Bug: T133748
Change-Id: I08013aa3aaf7c2a077765437180b7d845ebb4dce