Simple upload feature for uploadable fields.
Instead of going for the Upload window that current uploadable field
provides this feature enables a simple upload dialog with provision to
preview the image and change or remove it.

This feature currently only supports fields with a single value (image)

Note: This feature is disabled by default and can be turned on by
setting $wgPageFormsSimpleUpload to true.

Change-Id: Icdd76c1e826cb6e7286669c283981c5e7aee5e1a