Remove meaningless comments repeating the class name twice
Remove all comments stating the "File holding the … class" as well as
"The … class."

These comments do nothing but repeating the class name. The generated
documentation will state that a file named (for example) Foo.php "holds
the Foo class". In an other place the documentation will state that the
class "Foo" is "the Foo class".

Such comments are worse than having no comments, because I need to read
them to understand they do not hold any additional information, and do
not explain anything. Removing these comments makes room for actual
improvements in the future.

Note there are actual mistakes in some of these comments, e.g. because
the class got renamed, but the comment not updated.

Change-Id: I54ad9ac68684f558228f1b646f97da2c14674ca8