Remove field_is_array API param
If the API needs to know whether a Cargo field holds a list of values or not, it gets that information directly from the relevant DB table. Before, this was set based on whether the form field doing the querying held a list or not - but the field doing the querying and the field being queried are not necessarily the same, so this heuristic doesn't always work.

Change-Id: I7441bb9ed745c3f4629c79ad1a440bf1ceccb723