Don't stop spinner before form rendering is completed
With the fix provided for T216805, the spinner is removed too quickly
and the form can move around quite a bit after the spinner is removed.

Moving the end of time timout function and reducing the length of the
timeout appears to fix both the spinner problem and continue to allow
the date picker to work.

Bug: T216805
Change-Id: I65f14a1e41057d1cf61f546e8bf265ae84b8ad52