Only normalize longitude when setting form value, not for marker
If we normalize the position of the marker to be between ±180°
then it won't appear where the user has clicked. Instead, we
leave Leaflet displaying the larger value, but normalize the
value that's put into the form field.

This is a correction of I1831a9758b1063fed46057b2a9d47094d54d84ae

Bug: T236604
Change-Id: I0a332e6840d1cf90b864e00cd05c91da7e091bae