Fix an UI bug in type=tokens when double-clicking near Delete button
There is a visual bug when double-clicking extremely close to
the Delete button within the token (in "type=tokens" input).

A token will become ruined (replaced with an empty square with
nothing but the delete button).

This is caused by a wrong HTML element receiving the event,
due to area below/near the Delete button not belonging to <div>
that contains the text (visible label) of this token.

Proposed change makes only double-clicks on the token itself
open the editing, not below the Delete button, etc.
(it wasn't safe for user to click there anyway, as the user
could have accidentally deleted the needed token)

Change-Id: I14e8b72399bb19c88721e2883d2136e68f4fb575