Do not expect the Parser by reference
There is no reason to have these functions *expect* a reference. The
only reason for this is if one of these functions wants to *replace*
the parser with another one. This should never happen.

In old PHP versions we don't use any more the extra & was sometimes
helpful to avoid PHP doing expensive clones of large objects like the
Parser, in this case. Modern PHP does not have this problem any more.

Change-Id: I541903535766d01af98026d33a935664109da16a