Fix The Appearance of Checkboxes
In checkboxes input type, the labels were closer to the checkbox ahead
of them than to their own. The change would be in 'skins/PageForms.css'
that needs to identify a new selector.
This Selector must depend on the `.ltr` or `rtl` class for the body.
This assign a margin around the label, 20px after the Label.
Also, there was another issue is solved, that could be a checkbox at the
end of one line and its label at the beginning of the next line. it
is solved by assigning `display: inline-block` for the label.

Bug: T217517
Change-Id: Ib8487b5772090d465595b49d6f3cb3dd82db5d11