Fix the problem of Label clicking does not affect checkboxes
In Multi Instance Templates, when adding a new template, clicking on the
label does not check/uncheck the checkbox.
The problem was because of the unused "for" attribute.
The fix of the problem depends on Removing the function which replaces
the value of for' as it always returns undefined value.

Bug: T217516
Change-Id: I11b6a1b32de5adb5a030c2b31f53feb66cc4bda9