Add parsing to "Values from namespace/category/concept"
The parsing with parameter "values=" works well, you can write
"values={{ucfirst:a,b,c}}". It doesn't with paremeters "values from
category=", "values from namespace=" or "values from concept".
The problem was caused of not parsing the parameter at all. So, adding
statement to parse $valuesSource would be good.
The Parsing of "values from category=" should be before calling
the function `ucfirst` that depends on the value of $wgCapitalLinks.

Bug: T246715
Change-Id: Ia1a2005e935238e8bef4bc7f641bfd13e6ae308e