Replace the last comma in list from #arraymap with a conjunction
When using an arraymap to present a list of items they are separated by
a comma, even the last item. It would be more realistic to make the last
comma another thing (i.e. "and" or "or").
At the function renderArrayMap, the return is changed from a simple
implode() to a for loop that check if it is the last item, then add the
conjunction that the user wrote.

Bug: T92870
Change-Id: I04c99c5aaf7f5532ae12d8bcccd28959e7c046b9