Turn the "...with autocomplete" inputs into aliases for combobox, tokens
As the Library jQuery UI is deprecated, the input types "text with
autocomplete" and "textarea with autocomplete" are converted into
aliases for "tokens" and "combobox" who are calling Select2.js Library.
Both of the "... with autocomplete" inputs can be used for either single
option or multiple options, depending on whether it is a list.
If it is a list, then, convert it into Tokens input, other wise, convert
it into Combobox.
This Condition is Handled by the function getAlias(). The function is
important because we may call the static functions without calling the
constructor of the Class, which assigns the static $alias attribute.

Bug: T248505
Change-Id: Icbcb9389b43f5d4122b2710b173d240f300f3b2d