Make the #forminput use Select2 for autocomplete
Since the jQuery UI is deprecated, the #forminput needed to be handled,
to get rid of using it. That needed calling Select2, and since the
input #forminput uses single value, it calls the combobox.
The combobox has an arrow, that maybe strange here, so, hiding it using
CSS can help

Bug: T248505
Change-Id: Ideeb550e89e5d247d3f1c59a81a4c3979ff7e28a