Make the Datetimepicker and Datepicker use OOUI
The inputs "datetimepicker" and "datepicker" used to use
jQuery UI, which is deprecated. So they are now calling the OOUI
widgets.
There is a difference in the format of the date, which is handled
in the function getConvertedFormat().
Most parameters also have a differnce in the format or their names,
and they get handled in the function getOptions().

Bug: T248505
Change-Id: I86cdf968daf8daf0a0698536ef47f993f23cc9d4