Remove use of jQuery UI CSS for combobox
The "ui-widget" class name was in place for a long time for the combobox input, even though it was switched from jQuery UI to Select2 in 2014; but somehow the recent JS changes caused the specific (and unwanted) CSS for that class to start showing up again. Thus, the class name is changing now.

Change-Id: Ib4a2ff3d01c70961abdf91a4f49bbeded2141cfd