Replace jsGrid with jExcel for spreadsheets
This change is heavily based on the patch at https://gerrit.wikimedia.org/r/c/mediawiki/extensions/PageForms/+/620672 by Amr El-Absy, which was part of the Google Summer of Code but was unfortunately uncompleted at the end of the summer. It is part of the overall effort to remove the use of jQuery UI from Page Forms (jsGrid uses jQuery UI). jExcel has some advantages over jsGrid anyway.

Change-Id: Ice7d4e5c5feaf9384c0f324a53979697a2a074d7