Fixes broken address input for map input types
The migration to OOUI\TextInputWidget at PF_OpenLayersInput.php did break the
geolocation address input (`.pfAddressInput`). The OOUI wraps the actual input
tag into a container and add the classes to the container, not to the input, in
result the JS code is unable to find the input same as it unable to correctly
handle key events on a wrapper.