Read size parameter of #forminput
Read the given size and use it for the input element. I guess this is
something weird from the perspective of OOUI. In the future, we could
mark the size parameter as deprecated and introduce a new parameter, for
example 'width', with usage of '{{#forminput:…|width=100%}}'.

Also remove the default value for size (formerly 25), and let OOUI set
the default.

Change-Id: Iebba8ebde62642ef03e5e895af3fa5eb5a87e331