Add PFValuesUtils::getMaxValuesToRetrieve()
The global variable $wgPageFormsMaxAutocompleteValues was being used for two things: a maximum number of values for autocompletion, and a maximum number of values for non-autocomplete inputs, like radiobuttons. This patch gets the variable to only be used for *non*-autocompletion - which is incorrect, of course, given its name, but this is the better course of action, given that its default value is 1,000, and not, say, 20. Hopefully this is just an intermediate step.

Change-Id: Idf2d984b21df6877f9fd9d45fb99e14d6f3e972b