Remove fancyBox JS library, replace with built-in popup JS
fancyBox was originally available under GPL, and the version within Page Forms is GPL, but at some point (version 4, maybe) the library became non-open-source, so just to be safe, it is being removed from Page Forms. Thankfully, the built-in JavaScript for handling popup forms seems to handle file upload popups well, so this change actually ends up being an improvement, in terms of both code simplicity and UI consistency.

There is one other feature in Page Forms that makes use of fancyBox: the popup forms used within "display=calendar". This feature was not updated, and is now broken. "display=calendar" at the moment seems to get little to no usage, so this is probably not a big deal, but it still should be fixed.

Change-Id: I455953a715026e92f9c320306a1ed23423da79f3