Use RIGOR_QUICK for authorization checks when rendering formedit tab
PFFormEditAction::displayTab() currently implicitly uses RIGOR_SECURE
for checking whether the acting user is able to perform the `formedit`
action for the purposes of rendering the corresponding tab. In a setup
with multiple DB servers, this causes a primary DB query on pageviews,
which is a performance and resiliency concern. Use RIGOR_QUICK instead
- this is also what core skin logic uses to determine whether to render
various tabs.

Bug: T338421
Change-Id: I741948c5e18266e570f9c3bd9a85b0e95a498355