Fix for handling "property=" parameter
Some of the Cargo + SMW code was moved up, to before the setPossibleValues() call, in PFFormField::newFromFormFieldTag(). The Cargo code seems to work fine in either location, but the SMW code was failing, because the call to $f->template_field->setSemanticProperty() was happening *after* the semantic property was being used to set the possible values.

Bug: T338753
Change-Id: I0820b847835c89151a7ae848ac715c3aefe85dc6