PFFormLinker: Use PageProps for fetching default form name
getDefaultForm() can be passed a nonexistent title, in which case it
will issue a query to the page_props table with pp_page = 0 to attempt
to fetch the default form name. It is however possible for invalid
entries with pp_page = 0 to sometimes get stuck in the page_props table
- if these include the PFDefaultForm property, it suddenly becomes the
default form for any page that does not have one explicitly set.

So, use the PageProps service to fetch the default form name instead, as
it is both simpler and also ensures that no results will be returned if
the given title does not exist.

Bug: T341945
Change-Id: I89df79426d26c79bf5cabfcaca21cdb79899a7fd