Fix for f4d7212f0bab for default input types in Special:CreateForm
The dropdown value for the default input type in each case is supposed to start with a ".", because there's special handling for that with both the JS code that sets the allowed parameters on select, and the PHP code that creates the form on submit. This value was set to blank at some point, which fixed the PHP part but not the JS part.

Change-Id: I4de8b5572fba16615ce1f7e0fa5bab263cd04053