Set context title before previewing
Setting the title of the `$this->mOptions['target']` as the context
title of the EditPage object `$editor`. Also storing it as its own
variable because WritePageData already uses it in the lines above.

This will fix compatibility with the CodeEditor extension which requires
the EditPage object being passed to showEditForm:initial hook to have a
context title set or it will throw an exception (see commit 212477a).

Adding spaces next to parentheses on the new line 1038 as per the
Wikimedia code-style.

Bug: T352623
Change-Id: Ib2bf206317ba6310061b3741b17d6bdda2ff8db5