Update Sortable.js to remove local modifications
Update the Sortable.js file because it was failing verification
due to local modifications. Those were done in order to make it
load globally and so this change also requires the
ext.pageforms.sortable module in the three places it's used.

And add a `foreign-resources` Composer script
in Composer to make it easier to run manageForeignResources
as e.g. `composer foreign verify`.

Bug: T398898
Change-Id: Icccd4b03796ccbfe163ffcab072f8a3482916512