Replace UploadWindow iframe with core's Upload dialog
Use the mw.Upload.Dialog of MediaWiki core to upload files for
input types with `uploadable` set. This removes the need for the
custom UploadWindow special page and associated form.

Also handle when the upload links are in a multiple instance
template and added dynamically with a new input ID.

Bug: T402155
Change-Id: I50bc368c11a19b0cf6d35ba403fbf10447c309ae