Remove 3 unnecessary eslint rule warnings
It is not clear why these are promoted to warnings. They are
not included in the default MediaWiki rules, and PageForms
has quite a few transgressions anyway so it doesn't seem like
anyone's listening to the warnings.

Getting rid of them makes the other (useful) warnings easier to
read and so more likely to be acted on.

Change-Id: I6aa7af65bfd9bf16cd23e64b07c63c5639419d0c