Revert "Use namespaced PHP classes"
This reverts commit 205eee4770dc5c535545c2eb232a7ad3d43829ea.

Reason for revert: Sorry; I try to keep compatibility with my extensions for at least the last two years of MediaWiki versions, if possible, which means that removing support for pre-1.43 versions should ideally wait another year. I should have given this patch a "-1", to indicate my view. This does look like a very helpful change, and I'm looking forward to merging in this, or something very like it, in the future.

Change-Id: I4dcd97ed62cf709885893cf1d044e5a2793fbb75